Fix build production
