--- web/src/app/api/db-test/route.js
@@ -1,11 +1,0 @@
-import pool from "@/lib/users/db"
-
-export async function GET() {
-  try {
-    const result = await pool.query("SELECT NOW()")
-    return Response.json({ success: true, time: result.rows[0].now })
-  } catch (error) {
-    console.error("DB connection error:", error)
-    return new Response("Database error", { status: 500 })
-  }
-}
